// src/api/models/ClienteInternoRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClienteInternoRequest = {
    nombres: string;
    apellidos: string;
    genero: string;
    fechaNacimiento: string;
    telefono?: string | null;
    correoElectronico?: string | null;
    rol: string;
    username: string;
    password: string;
    tipoDocumento: string;
    numeroIdentificacion: string;
};